+++ seed/template/all/app/components/app/_app.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+    // moduleId: module.id,
+    selector: 'my-app',
+    styles: [`{%= include('app.component.css') %}`],
+    template: `{%= include('app.component.html') %}`
+})
+
+export class AppComponent {
+    title = 'Tour of Heroes';
+}