+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 - [Spanish](#spanish)
 
 ## English
+### v.0.3.0
+    - Updated to Angular2 v2.1.1
+    - Updated to generator-speedseed v.0.22.0
+    - Updated to speedseed v.0.18.0
+
 ### v.0.2.2
     - Updated to generator-speedseed v.0.21.0
     - Updated to speedseed v.0.17.0
@@ -15,6 +20,11 @@
     - Init
 
 ## Spanish
+### v.0.3.0
+    - Actualizado a Angular2 v2.1.1
+    - Actualizado a generator-speedseed v.0.22.0
+    - Actualizado a speedseed v.0.18.0
+
 ### v.0.2.2
     - Actualizado a generator-speedseed v.0.21.0
     - Actualizado a speedseed v.0.17.0